cleaned up the code
