+++ app/components/Favorites/Favorites.js
@@ -2,27 +2,21 @@ import React from 'react'
 import MovieCard from '../MovieCard/MovieCard'
 import styles from  './Favorites.css'
 
-
 const Favorites = ({movies,favorites,handleFavorites}) => {
   let list = movies.filter(movie => {
   return favorites.includes(movie.id)
 })
-console.log(list)
 
 let renderCard = list.length ?
   list.map((movie,i)=>{
-
-  return (
-    <MovieCard
-    handleFavorites={handleFavorites}
-    key={i}
-    index={movie.id}
-    movieInfo={movie}/>
-  )
-})
-  :
-  <div className="no-favorites-container">No favorite movies? Come on...</div>
-
+    return (
+      <MovieCard
+      handleFavorites={handleFavorites}
+      key={i}
+      index={movie.id}
+      movieInfo={movie}/>
+    )
+  }) : <div className="no-favorites-container">No favorite movies? Come on...</div>
 
   return(
     <section className="favorites-list">
