+++ app/components/Login/Login.js
@@ -1,7 +1,6 @@
 import React, {Component} from 'react'
 import styles from  './Login.css'
 
-
 class Login extends Component {
   constructor(props){
     super(props)
@@ -14,57 +13,54 @@ class Login extends Component {
   }
 
   logInfo() {
-     if(this.state.password.value === this.state.retypedPassword.value) {
-       if(this.state.name !== '' && this.state.email !== '') {
-         fetch("api/users/new", {
-           method: "POST",
-           headers: {"Content-Type": "application/json"},
-           body: JSON.stringify({name: this.state.name,
-                  email: this.state.email,
-                  password: this.state.password})
-         }).then((response) => {
-           console.log(response)
-           this.props.handleUser({name:this.state.name,email:this.state.email})
-         })
-       }
-     }
-}
-render(){
+    if(this.state.password.value === this.state.retypedPassword.value) {
+      if(this.state.name !== '' && this.state.email !== '') {
+        fetch("api/users/new", {
+          method: "POST",
+          headers: {"Content-Type": "application/json"},
+          body: JSON.stringify({name: this.state.name,
+                email: this.state.email,
+                password: this.state.password})})
+        .then((response) => {
+          this.props.handleUser({name:this.state.name,email:this.state.email})})
+      }
+    }
+  }
 
-  return(
-    <section className = "login-page">
-    <input
-    placeholder = "name"
-    className = "login-name"
-     value = {this.state.name}
-     onChange = {(e)=>{this.setState({name:e.target.value})}}
-     />
-
-    <input
-    placeholder = "email"
-    className = "login-email"
-     value = {this.state.email}
-     onChange = {(e)=>{this.setState({email:e.target.value})}}
-     />
-    <input
-      placeholder = "password"
-      className = "login-password"
-      value = {this.state.password}
-      onChange = {(e)=>{this.setState({password:e.target.value})}}
-      />
-      <input
-        placeholder = "retype password"
-       className = "login-retypedPassword"
-        value = {this.state.retypedPassword}
-        onChange = {(e)=>{this.setState({retypedPassword:e.target.value})}}
+  render(){
+    return(
+      <section className = "login-page">
+        <input
+        placeholder = "name"
+        className = "login-name"
+         value = {this.state.name}
+         onChange = {(e)=>{this.setState({name:e.target.value})}}
+        />
+        <input
+        placeholder = "email"
+        className = "login-email"
+         value = {this.state.email}
+         onChange = {(e)=>{this.setState({email:e.target.value})}}
+         />
+        <input
+          placeholder = "password"
+          className = "login-password"
+          value = {this.state.password}
+          onChange = {(e)=>{this.setState({password:e.target.value})}}
+          />
+        <input
+          placeholder = "retype password"
+         className = "login-retypedPassword"
+          value = {this.state.retypedPassword}
+          onChange = {(e)=>{this.setState({retypedPassword:e.target.value})}}
         />
         <button
         className="login-submit"
          onClick={()=>{this.logInfo()}}>
         Submit
         </button>
-    </section>
-  )
+      </section>
+    )
   }
 }
