+++ app/components/MovieList/MovieList.js
@@ -3,11 +3,8 @@ import MovieCard from '../MovieCard/MovieCard'
 import styles from  './MovieList.css'
 
 const MovieList = (props) => {
-
-
-    let renderCard = props.movies ?
+  let renderCard = props.movies ?
     props.movies.map((movie,i)=>{
-
       return (
         <MovieCard
         handleFavorites={props.handleFavorites}
@@ -15,17 +12,11 @@ const MovieList = (props) => {
         index={movie.id}
         movieInfo={movie}/>
       )
-})
-      :
-      <div>loading</div>
-
+    }) : <div>loading</div>
 
   return(
     <section className="movie-list">
-    {
-      renderCard
-    }
-
+    {renderCard}
     </section>
   )
 }
