+++ app/components/MovieList/MovieListContainer.js
@@ -14,6 +14,4 @@ const mapDispatchToProps = (dispatch) => {
   }
 }
 
-
-
 export default connect(mapStateToProps,mapDispatchToProps)(MovieList)
