+++ app/components/NavBar/NavBar.css
@@ -1,8 +1,8 @@
-.navBar{
+.navBar {
   margin-bottom: 10px;
 }
 
-.nav-link{
+.nav-link {
   background-color:#fff;
   padding: 5px;
   font-weight: 100;
