+++ app/components/NavBar/NavBar.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from 'react'
 import { NavLink } from 'react-router-dom'
 import styles from  './NavBar.css'
 
@@ -9,5 +9,5 @@ export const NavBar = () => {
       <NavLink className="nav-link" activeClassName='selected' to={'/favorites'}> Favorites </NavLink>
       <NavLink className="nav-link" activeClassName='selected' to={'/login'}> Login </NavLink>
     </div>
-  );
+  )
 }
